% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansiex.R
\name{ansi_regex}
\alias{ansi_regex}
\title{Perl compatible regular expression that matches ANSI escape
sequences}
\usage{
ansi_regex()
}
\value{
String scalar, the regular expression.
}
\description{
Don't forget to use \code{perl = TRUE} when using this with \code{\link[=grepl]{grepl()}} and
friends.
}
\seealso{
Other low level ANSI functions: 
\code{\link{ansi_has_any}()},
\code{\link{ansi_hide_cursor}()},
\code{\link{ansi_string}()},
\code{\link{ansi_strip}()}
}
\concept{low level ANSI functions}
