% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansiex.R
\name{ansi_simplify}
\alias{ansi_simplify}
\title{Simplify ANSI styling tags}
\usage{
ansi_simplify(x)
}
\arguments{
\item{x}{Input string}
}
\value{
Simplified \code{ansi_string} vector.
}
\description{
It creates an equivalent, but possibly shorter ANSI styled string, by
removing duplicate and empty tags.
}
