% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansiex.R
\name{ansi_string}
\alias{ansi_string}
\title{Labels a character vector as containing ANSI control codes.}
\usage{
ansi_string(x)
}
\arguments{
\item{x}{A character vector or something that can be
coerced into one.}
}
\value{
A \code{cli_ansi_string} object, a subclass of
\code{character}, with the same length and contents
as \code{x}.
}
\description{
This function sets the class of its argument, activating
ANSI-string-specific methods such as for printing.
}
\seealso{
Other low level ANSI functions: 
\code{\link{ansi_has_any}()},
\code{\link{ansi_hide_cursor}()},
\code{\link{ansi_regex}()},
\code{\link{ansi_strip}()}
}
\concept{low level ANSI functions}
