% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonex.R
\name{ansi_trim_ws}
\alias{ansi_trim_ws}
\title{Remove leading and/or trailing whitespace from an ANSI string}
\usage{
ansi_trim_ws(x, which = c("both", "left", "right"))
}
\arguments{
\item{x}{ANSI string vector.}

\item{which}{Whether to remove leading or trailing whitespace or both.}
}
\value{
ANSI string, with the whitespace removed.
}
\description{
This function is similar to \code{\link[base:trimws]{base::trimws()}} but works on ANSI strings,
and keeps color and other styling.
}
\examples{
trimws(paste0("   ", col_red("I am red"), "   "))
ansi_trim_ws(paste0("   ", col_red("I am red"), "   "))
trimws(col_red("   I am red   "))
ansi_trim_ws(col_red("   I am red   "))
}
