% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-component-tree.R
\name{apply_theme_to_style}
\alias{apply_theme_to_style}
\title{Apply a theme to a component style}
\usage{
apply_theme_to_style(cpt_styles, thm_styles)
}
\arguments{
\item{cpt_styles}{Component styles, a named list.}

\item{thm_styles}{Theme styles, a named list.}
}
\description{
This is a helper function of \code{\link[=theme_component_tree]{theme_component_tree()}}, to merge
component styles and a theme.
}
\details{
The component style takes precedence over the theme, except
for the \code{class-map} style, which is merged elementwise, again, with
the component style having a higher precedence.
}
\keyword{internal}
