% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{boxes}
\alias{boxes}
\alias{boxes-package}
\title{Draw Boxes in the 'R' Console}
\description{
It integrates well with 'crayon' for coloring the boxes or their
content, and 'clisymbols' to include Unicode characters in the boxes.
It can also draw rules to make headers or footers in console output.
}
