% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat-helpers.R
\name{cat_line}
\alias{cat_line}
\alias{cat_bullet}
\alias{cat_boxx}
\alias{cat_rule}
\alias{cat_print}
\title{\code{cat()} helpers}
\usage{
cat_line(..., file = stdout())

cat_bullet(..., file = stdout())

cat_boxx(..., file = stdout())

cat_rule(..., file = stdout())

cat_print(x, file = stdout())
}
\arguments{
\item{...}{For \code{cat_line()} and \code{cat_bullet()}, passed on to \code{\link[=cat]{cat()}} and
pasted together. For \code{cat_rule()} and \code{cat_boxx()} passed on to
\code{\link[=rule]{rule()}} and \code{\link[=boxx]{boxx()}} respectively.}

\item{file}{Output destination. Defaults to standard output.}

\item{x}{An object to print.}
}
\description{
These helpers provide useful wrappers around \code{\link[=cat]{cat()}}: most importantly
they all set \code{sep = ""}, and \code{cat_line()} automatically adds a newline.
}
