% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-package.R
\docType{package}
\name{cli-package}
\alias{cli-package}
\title{cli: Helpers for Developing Command Line Interfaces}
\description{
A suite of tools to build attractive command line interfaces ('CLIs'), from semantic elements: headings, lists, alerts, paragraphs, etc. Supports custom themes via a 'CSS'-like language. It also contains a number of lower level 'CLI' elements: rules, boxes, trees, and 'Unicode' symbols with 'ASCII' alternatives. It support ANSI colors and text styles as well.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cli.r-lib.org}
  \item \url{https://github.com/r-lib/cli}
  \item Report bugs at \url{https://github.com/r-lib/cli/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Other contributors:
\itemize{
  \item Hadley Wickham [contributor]
  \item Kirill Müller [contributor]
  \item Salim Brüggemann \email{salim-b@pm.me} (\href{https://orcid.org/0000-0002-5329-5987}{ORCID}) [contributor]
  \item RStudio [copyright holder, funder]
}

}
\keyword{internal}
