% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{cli-themes}
\alias{cli-themes}
\title{Themes for command line interfaces (CLIs)}
\description{
A cli theme is a named list. The names of the list elements are
CSS selectors that define whith CLI elements each style element applies
to. See \url{https://www.w3schools.com/cssref/css_selectors.asp} for a
reference on CSS selectors. We will also show some examples below.
}
\details{
Themes can come from three sources:
\itemize{
\item The builtin theme of the cli package.
\item The theme from the \code{cli.theme} option. (At the time when the
\code{cli_class} object is created.)
\item Themes added with he \code{add_theme()} method.
}
}
\section{Examples}{

Color of headers, that are only active in paragraphs with an
'output' class:\preformatted{list(
  "par.output h1" = list("background-color" = "red", color = "#e0e0e0"),
  "par.output h2" = list("background-color" = "orange", color = "#e0e0e0"),
  "par.output h3" = list("background-color" = "blue", color = "#e0e0e0")
)
}

Create a custom alert type:\preformatted{list(
  ".alert-start" = list(before = symbol$play),
  ".alert-stop"  = list(before = symbol$stop)
)
}
}

