% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli}
\alias{cli}
\title{Compose multiple cli functions}
\usage{
cli(expr)
}
\arguments{
\item{expr}{Expression that contains \verb{cli_*} calls. Their output is
collected and sent as a single message.}
}
\value{
Nothing.
}
\description{
\code{cli()} will record all \verb{cli_*} calls in \code{expr}, and emit them together
in a single message. This is useful if you want to built a larger
piece of output from multiple \verb{cli_*} calls.
}
\details{
Use this function to build a more complex piece of CLI that would not
make sense to show in pieces.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cli(\{
  cli_h1("Title")
  cli_h2("Subtitle")
  cli_ul(c("this", "that", "end"))
\})
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#>                                                                                 
#> <span style="color: #2AA198;">──</span> <span style="font-weight: bold;">Title</span> <span style="color: #2AA198;">─────────────────────────────────────────────────────────────</span>          
#>                                                                                 
#> ── <span style="font-weight: bold;">Subtitle</span> ──                                                                  
#>                                                                                 
#> • this                                                                          
#> • that                                                                          
#> • end                                                                           
</pre></div>
}}
}
