% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_blockquote}
\alias{cli_blockquote}
\title{CLI block quote}
\usage{
cli_blockquote(
  quote,
  citation = NULL,
  id = NULL,
  class = NULL,
  .envir = parent.frame()
)
}
\arguments{
\item{quote}{Text of the quotation.}

\item{citation}{Source of the quotation, typically a link or the name
of a person.}

\item{id}{Element id, a string. If \code{NULL}, then a new id is generated
and returned.}

\item{class}{Class name, sting. Can be used in themes.}

\item{.envir}{Environment to evaluate the glue expressions in. It is
also used to auto-close the container if \code{.auto_close} is \code{TRUE}.}
}
\description{
A section that is quoted from another source. It is typically indented.
}
