% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\docType{data}
\name{cli_class}
\alias{cli_class}
\title{Create a command line interface}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
cli_class
}
\description{
TODO
}
\keyword{datasets}
