% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_par}
\alias{cli_par}
\title{CLI paragraph}
\usage{
cli_par(id = NULL, class = NULL, .auto_close = TRUE, .envir = parent.frame())
}
\arguments{
\item{id}{Element id, a string. If \code{NULL}, then a new id is generated
and returned.}

\item{class}{Class name, sting. Can be used in themes.}

\item{.auto_close}{Whether to close the container, when the calling
function finishes (or \code{.envir} is removed, if specified).}

\item{.envir}{Environment to evaluate the glue expressions in. It is
also used to auto-close the container if \code{.auto_close} is \code{TRUE}.}
}
\value{
The id of the new container element, invisibly.
}
\description{
See \link{containers}.
}
\examples{
id <- cli_par()
cli_text("First paragraph")
cli_end(id)
id <- cli_par()
cli_text("Second paragraph")
cli_end(id)
}
