% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_progress_bar}
\alias{cli_progress_bar}
\title{CLI progress bar}
\usage{
cli_progress_bar(...)
}
\arguments{
\item{...}{All arguments are passed to the constuctor of the
\link[progress:progress_bar]{progress::progress_bar} class.}
}
\value{
A remote progress bar object that can be used the same way
as \link[progress:progress_bar]{progress::progress_bar}, see examples below.
}
\description{
A progress bar using the progress package
}
\examples{
{
  p <- cli_progress_bar(total = 10)
  cli_alert_info("Starting computation")
  for (i in 1:10) { p$tick(); Sys.sleep(0.2) }
  cli_alert_success("Done")
}
}
