% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-client.R
\name{cli_progress_output}
\alias{cli_progress_output}
\title{Add text output to a progress bar}
\usage{
cli_progress_output(text, id = NULL, .envir = parent.frame())
}
\arguments{
\item{text}{Text to output. It is formatted via \code{\link[=cli_text]{cli_text()}}.}

\item{id}{Progress bar id. The default is the current progress bar.}

\item{.envir}{Environment to use for glue interpolation of \code{text}.}
}
\value{
\code{TRUE}, always.
}
\description{
The text is calculated via \code{\link[=cli_text]{cli_text()}}, so all cli features can be
used here, including progress variables.
}
\details{
The text is passed to the progress handler(s), that may or may not be
able to print it.
}
