% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep.R
\name{cli_sitrep}
\alias{cli_sitrep}
\title{cli situation report}
\usage{
cli_sitrep()
}
\value{
Named list with entries listed above. It has a \code{cli_sitrep}
class, with a \code{print()} and \code{format()} method.
}
\description{
Contains currenty:
\itemize{
\item \code{cli_unicode_option}: whether the \code{cli.unicode} option is set and its
value. See \code{\link[=is_utf8_output]{is_utf8_output()}}.
\item \code{symbol_charset}: the selected character set for \link{symbol}, UTF-8,
Windows, or ASCII.
\item \code{console_utf8}: whether the console supports UTF-8. See
\code{\link[base:l10n_info]{base::l10n_info()}}.
\item \code{latex_active}: whether we are inside knitr, creating a LaTeX
document.
\item \code{num_colors}: number of ANSI colors. See \code{\link[=num_ansi_colors]{num_ansi_colors()}}.
\item \code{console_with}: detected console width.
}
}
\examples{
cli_sitrep()
}
