% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-bar.R
\name{cli_status_update}
\alias{cli_status_update}
\title{Update the status bar}
\usage{
cli_status_update(..., id = NULL, .envir = parent.frame())
}
\arguments{
\item{...}{Text to update the status bar with.}

\item{id}{Id of the status bar to update. Defaults to the current
status bar container.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\value{
Id of the status bar container.
}
\description{
Update the status bar
}
\seealso{
Other status bar: 
\code{\link{cli_status_clear}()},
\code{\link{cli_status}()}
}
\concept{status bar}
