% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cli_table}
\alias{cli_table}
\title{Create a table}
\usage{
cli_table(
  ...,
  data = NULL,
  th = NULL,
  id = NULL,
  class = NULL,
  .close = TRUE,
  .auto_close = TRUE,
  .envir = parent.frame()
)
}
\arguments{
\item{...}{The rows of the table, character vectors.}

\item{data}{if not \code{NULL}, then a data frame to attach to the
search path while creating the table.}

\item{th}{A header row, optionally.}

\item{id}{Element id, a string. If \code{NULL}, then a new id is generated
and returned.}

\item{class}{Class name, sting. Can be used in themes.}

\item{.close}{Whether to close the table. If it is not close then
more lines can be added to it.}

\item{.auto_close}{Whether to close the container, when the calling
function finishes (or \code{.envir} is removed, if specified).}

\item{.envir}{Environment to evaluate the glue expressions in. It is
also used to auto-close the container if \code{.auto_close} is \code{TRUE}.}
}
\value{
The id of the table container, invisibly.
}
\description{
Create a table
}
\seealso{
Other tables: 
\code{\link{cli_th}()},
\code{\link{cli_tr}()}
}
\concept{tables}
