% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{cli_th}
\alias{cli_th}
\title{Add a header row to a table}
\usage{
cli_th(..., data = NULL, id = NULL, class = NULL, .envir = parent.frame())
}
\arguments{
\item{...}{One or more character vectors, the headers row(s).}

\item{data}{if not \code{NULL}, then a data frame to attach to the
search path while creating the table.}

\item{id}{Element id, a string. If \code{NULL}, then a new id is generated
and returned.}

\item{class}{Class name, sting. Can be used in themes.}

\item{.envir}{Environment to evaluate the glue expressions in. It is
also used to auto-close the container if \code{.auto_close} is \code{TRUE}.}
}
\value{
The id of the header container, invisibly.
}
\description{
Add a header row to a table
}
\seealso{
Other tables: 
\code{\link{cli_table}()},
\code{\link{cli_tr}()}
}
\concept{tables}
