% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{cli_with_progress_ticks}
\alias{cli_with_progress_ticks}
\title{Evaluate an expression with forcing all progress bar updates enabled}
\usage{
cli_with_progress_ticks(expr)
}
\arguments{
\item{expr}{Expression to evaluate.}
}
\value{
The value of \code{expr}.
}
\description{
Evaluates \code{expr} while redrawing the progress bar for every progress
bar update. This function is meant to be used in test cases, when
specifically testing the progress bar output.
}
