% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-tree.R
\name{component-trees}
\alias{component-trees}
\alias{component-tree}
\alias{component_tree}
\alias{component_trees}
\title{Component trees}
\description{
Component trees are used to calculate the final appearance of a
component and its subcomponents.
}
\details{
Rendering a component has the following steps:
\enumerate{
\item Mapping with \code{\link[=map_component_tree]{map_component_tree()}}. It calculates the paths of the
component and its subcomponents. The paths are needed for theming.
\item Theming with \code{\link[=theme_component_tree]{theme_component_tree()}}. Styles from a theme are added
to the nodes of the component tree, according to the theme selectors.
\item Styling with \code{\link[=style_component_tree]{style_component_tree()}}. Styles are inherited
according to the style inheritance rules.
}
}
\seealso{
Other component trees: 
\code{\link{map_component_tree}()},
\code{\link{style_component_tree}()},
\code{\link{theme_component_tree}()}
}
\concept{component trees}
\keyword{internal}
