% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_div}
\alias{cpt_div}
\title{Create a generic block component}
\usage{
cpt_div(..., children = NULL, attr = NULL)
}
\arguments{
\item{...}{Subcomponents to add to the new component.}

\item{children}{A list of subcomponents to add to the new component,
they are added after the ones in \code{...}.}

\item{attr}{Attributes.}
}
\value{
A \code{div} component object.
}
\description{
Generic components are useful for styling, and for collecting
multiple components into a single entitity.
}
