% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_h1}
\alias{cpt_h1}
\alias{cpt_h2}
\alias{cpt_h3}
\title{Create headers}
\usage{
cpt_h1(text, attr = NULL, .envir = parent.frame())

cpt_h2(text, attr = NULL, .envir = parent.frame())

cpt_h3(text, attr = NULL, .envir = parent.frame())
}
\arguments{
\item{text}{Header text.}

\item{attr}{Attributes.}

\item{.envir}{Environment to evaluate interpolated expressions in.}
}
\value{
Header component.
}
\description{
Create headers
}
