% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_li}
\alias{cpt_li}
\alias{li}
\title{List item component}
\usage{
cpt_li(..., children = NULL, class = NULL, style = NULL, attr = NULL)

li(..., children = NULL, class = NULL, style = NULL, attr = NULL)
}
\arguments{
\item{...}{Child components.}

\item{children}{More child components, in a list.}

\item{class}{Classes.}

\item{style}{Styles.}

\item{attr}{Attributes.}
}
\value{
A list item component, with class \code{cli_component_li}.
}
\description{
Creates a list item component.
It can be used in \link[=cpt_ul]{unordered} or \link[=cpt_ol]{ordered} lists.
}
