% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_span}
\alias{cpt_span}
\alias{span}
\title{Create a generic inline component}
\usage{
cpt_span(..., children = NULL, class = NULL, style = NULL, attr = NULL)

span(..., children = NULL, class = NULL, style = NULL, attr = NULL)
}
\arguments{
\item{...}{Subcomponents to add to the new component.}

\item{children}{A list of subcomponents to add to the new component,
they are added after the ones in \code{...}.}

\item{class}{Classes.}

\item{style}{Styles.}

\item{attr}{Attributes.}
}
\value{
A \code{span} component object.
}
\description{
Generic components are useful for styling, and for collecting
multiple components into a single entitity.
}
