% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_text}
\alias{cpt_text}
\title{Create a \emph{text} component}
\usage{
cpt_text(..., .envir = parent.frame(), .call = sys.call())
}
\arguments{
\item{...}{The text to show, in character vectors, or other \emph{text}
components. Adjacent character vectors will be concatenated into a
single string. Newlines are \emph{not} preserved.}

\item{.envir}{Environment to evaluate the glue expressions in.}

\item{.call}{Call to pass down, for better error messages.}
}
\description{
Create a \emph{text} component
}
\concept{cli components}
