% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-text.R
\name{cpt_txt}
\alias{cpt_txt}
\title{Text component}
\usage{
cpt_txt(txt, .envir = parent.frame())
}
\arguments{
\item{txt}{Character vector. May contain interpreted \code{{}} literals.}

\item{.envir}{Environment to evaluate the interpreted literals in.}
}
\value{
A text component, with class \code{cli_component_text}.
}
\description{
Text component
}
