% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_ul}
\alias{cpt_ul}
\alias{ul}
\title{Unordered list component}
\usage{
cpt_ul(items, ..., attr = NULL, .envir = parent.frame())

ul(items, ..., attr = NULL, .envir = parent.frame())
}
\arguments{
\item{items, ...}{List items, a list or a a character vector. Elements
must be strings or \link[=cpt_li]{list item components}. Character strings
can have interpreted \code{{}} literals.}

\item{attr}{Attributes.}

\item{.envir}{Environment to evaluate the interpreted \code{{}} literals in
items in.}
}
\value{
An unordered list component, with class \code{cli_component_ul}.
}
\description{
It creates a bulleted list by default.
This can be customized with the \code{list-style-type} style.
}
