% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{cpt_verbatim}
\alias{cpt_verbatim}
\title{Create a \emph{verbatim} component}
\usage{
cpt_verbatim(..., .envir = parent.frame(), .call = sys.call())
}
\arguments{
\item{...}{TOFO}

\item{.envir}{TODO}

\item{.call}{TODO}
}
\description{
Create a \emph{verbatim} component
}
