% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{demo_spinners}
\alias{demo_spinners}
\title{Show a demo of some (by default all) spinners}
\usage{
demo_spinners(which = NULL)
}
\arguments{
\item{which}{Character vector, which spinners to demo.}
}
\description{
Each spinner is shown for about 2-3 seconds.
}
\examples{
\dontrun{
  demo_spinners(sample(list_spinners(), 10))
}
}
\seealso{
Other spinners: \code{\link{get_spinner}},
  \code{\link{list_spinners}}
}
\concept{spinners}
