% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{faq}
\alias{faq}
\title{Frequently Asked Questions}
\description{
Frequently Asked Questions
}
\details{
\subsection{My platform supports ANSI colors, why does cli not use them?}{

It is probably a mistake in cli’s ANSI support detection algorithm.
Please open an issue at \url{https://github.com/r-lib/cli/issues} and don’t
forget to tell us the details of your platform and terminal or GUI.
}

\subsection{How do I turn off ANSI colors and styles?}{

Set the \code{NO_COLOR} environment variable to a non-empty value. You can do
this in your \code{.Renviron} file (use \code{usethis::edit_r_environ()}).

If you want to do this for testthat tests, then consider using the 3rd
edition on testthat, which does turn off ANSI styling automatically
inside \code{test_that()}.
}
}
