% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keypress.R
\name{has_keypress_support}
\alias{has_keypress_support}
\title{Check if the current platform/terminal supports reading
single keys.}
\usage{
has_keypress_support()
}
\value{
Whether there is support for waiting for individual
keypressses.
}
\description{
Check if the current platform/terminal supports reading
single keys.
}
\details{
Supported platforms:
\itemize{
\item Terminals in Windows and Unix.
\item RStudio terminal.
}

Not supported:
\itemize{
\item RStudio (if not in the RStudio terminal).
\item R.app on macOS.
\item Rgui on Windows.
\item Emacs ESS.
\item Others.
}
}
\examples{
has_keypress_support()
}
\seealso{
Other keypress function: 
\code{\link{keypress}()}
}
\concept{keypress function}
