% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash_md5}
\alias{hash_md5}
\title{MD5 hash}
\usage{
hash_md5(x)
}
\arguments{
\item{x}{Character vector. If not a character vector, then
\code{\link[=as.character]{as.character()}} is used to try to coerce it into one. \code{NA} entries
will have an \code{NA} hash.}
}
\value{
Character vector of hexadecimal MD5 hashes.
}
\description{
Calculate the MD5 hash of each element of a character vector.
}
\examples{
hash_md5(c("foo", NA, "bar", ""))
}
\seealso{
\code{\link[tools:md5sum]{tools::md5sum()}} for a base R MD5 function that works on
files.

Other hash functions: 
\code{\link{hash_animal}()},
\code{\link{hash_emoji}()}
}
\concept{hash functions}
