% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{hash_sha256}
\alias{hash_sha256}
\alias{hash_raw_sha256}
\alias{hash_obj_sha256}
\alias{hash_file_sha256}
\title{SHA-256 hash}
\usage{
hash_sha256(x)

hash_raw_sha256(x)

hash_obj_sha256(x, serialize_version = 2)

hash_file_sha256(paths)
}
\arguments{
\item{x}{Character vector. If not a character vector, then
\code{\link[=as.character]{as.character()}} is used to try to coerce it into one. \code{NA} entries
will have an \code{NA} hash.}

\item{serialize_version}{Workspace format version to use, see
\code{\link[base:serialize]{base::serialize()}}.}
}
\value{
\code{hash_sha256()} returns aharacter vector of hexadecimal
SHA-256 hashes.

\code{hash_raw_sha256()} returns a character scalar.

\code{hash_obj_sha256()} returns a character scalar.

\code{hash_file_sha256()} returns a character vector of SHA-256
hashes.
}
\description{
Calculate the SHA-256 hash of each element of a character vector.
}
\details{
\code{hash_raw_sha256()} calculates the SHA-256 hash of the bytes
of a raw vector.

\code{hash_obj_sha256()} calculates the SHA-256 hash of an R
object. The object is serialized into a binary vector first.

\code{hash_file_sha256()} calculates the SHA-256 hash of one or
more files.
}
\examples{
hash_sha256(c("foo", NA, "bar", ""))
}
\seealso{
Other hash functions: 
\code{\link{hash_animal}()},
\code{\link{hash_emoji}()},
\code{\link{hash_md5}()}
}
\concept{hash functions}
