% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi-hyperlink.R
\name{hyperlink}
\alias{hyperlink}
\alias{ansi_has_hyperlink_support}
\alias{hyperlink_help}
\alias{hyperlink_vignette}
\alias{hyperlink_file}
\alias{hyperlink_viewer}
\title{Terminal Hyperlinks}
\usage{
hyperlink(text, url, params = NULL)

ansi_has_hyperlink_support()

hyperlink_help(text, topic, package = NULL)

hyperlink_vignette(text, topic, package = NULL)

hyperlink_file(text, file, line = NULL, col = NULL)

hyperlink_viewer(text, url)
}
\arguments{
\item{text}{Text to show. \code{text} and \code{url} are recycled to match their
length, via a \code{paste0()} call.}

\item{url}{URL to link to.}

\item{params}{A named character vector of additional parameters, or \code{NULL}.}

\item{topic}{topic for help page or vignette.}

\item{package}{package for help page or vignette.}

\item{line}{Line in the file.}

\item{col}{Column in the file.}
}
\value{
Styled \code{ansi_string} for \code{hyperlink()}.
Logical scalar for \code{ansi_has_hyperlink_support()}.
}
\description{
\code{hyperlink()} creates an ANSI hyperlink.
}
\details{
This function is currently experimental. In particular, many of the
\verb{ansi_*()} functions do not support it properly.

\code{ansi_has_hyperlink_support()} checks if the current \code{stdout()}
supports hyperlinks.

See also
\url{https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda}.
}
\examples{
cat("This is an", hyperlink("R", "https://r-project.org"), "link.\n")
ansi_has_hyperlink_support()
}
