% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-component-tree.R
\name{inherit_styles}
\alias{inherit_styles}
\title{Style inhertance rules}
\usage{
inherit_styles(parent, child, tag)
}
\arguments{
\item{parent}{Parent styles, a named list.}

\item{child}{Child styles, a named list.}

\item{tag}{Tag name of the \emph{child} node, a string.}
}
\value{
Merged styles, a named list.
}
\description{
All style inheritance rules are implemented here.
}
\details{
\itemize{
\item All styles in \code{inherited_styles()} are inherited.
\item \code{class-map} is inherited specially, in that the parent's map
is merged with the child's map, with the child's map taking
precedence.
\item Other styles are not inherited.
}
}
\keyword{internal}
