% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utf8.R
\name{is_utf8_output}
\alias{is_utf8_output}
\title{Whether cli is emitting UTF-8 characters}
\usage{
is_utf8_output()
}
\value{
Flag, whether cli uses UTF-8 characters.
}
\description{
UTF-8 cli characters can be turned on by setting the \code{cli.unicode}
option to \code{TRUE}. They can be turned off by setting if to \code{FALSE}.
If this option is not set, then \code{\link[base:l10n_info]{base::l10n_info()}} is used to detect
UTF-8 support.
}
