% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spinner.R
\name{list_spinners}
\alias{list_spinners}
\title{List all available spinners}
\usage{
list_spinners()
}
\value{
Character vector of all available spinner names.
}
\description{
List all available spinners
}
\examples{
list_spinners()
get_spinner(list_spinners()[1])
}
\seealso{
Other spinners: \code{\link{demo_spinners}},
  \code{\link{get_spinner}}
}
