% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component-tree.R
\name{map_component_tree}
\alias{map_component_tree}
\title{Map a component tree}
\usage{
map_component_tree(cpt, parent = list())
}
\arguments{
\item{cpt}{Component.}

\item{parent}{Parent components, if any. A list where each element
represents a parent node. Each element is a named list of elements:
\itemize{
\item \code{tag}: componnent tag,
\item \code{id}: component id, from the \code{id} attribute,
\item \code{class}: compoennt class or classes, from the \code{class} attribute.
}}
}
\value{
A mapped component tree, a \code{cli_component_tree} object.
}
\description{
We need to walk the subcomponents of the input component, and during the
walk we assign a \code{path} element to every node of the component tree.
}
\seealso{
Other component trees: 
\code{\link{component-trees}},
\code{\link{style_component_tree}()},
\code{\link{theme_component_tree}()}
}
\concept{component trees}
\keyword{internal}
