% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{match_selector}
\alias{match_selector}
\title{Match a selector to a container stack}
\usage{
match_selector(sels, cnts)
}
\arguments{
\item{sels}{A list of selector nodes.}

\item{cnts}{A list of container nodes.

The last selector in the list must match the last container, so we
do the matching from the back. This is because we use this function
to calculate the style of newly encountered containers.}
}
\description{
Match a selector to a container stack
}
