% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{match_selector_node}
\alias{match_selector_node}
\title{Match a selector node to a container}
\usage{
match_selector_node(node, cnt)
}
\arguments{
\item{node}{Selector node, as parsed by \code{parse_selector_node()}.}

\item{cnt}{Container node, has elements \code{tag}, \code{id}, \code{class}.

The selector node matches the container, if all these hold:
\itemize{
\item The id of the selector is missing or unique.
\item The tag of the selector is missing or unique.
\item The id of the container is missing or unique.
\item The tag of the container is unique.
\item If the selector specifies an id, it matches the id of the container.
\item If the selector specifies a tag, it matches the tag of the container.
\item If the selector specifies class names, the container has all these
classes.
}}
}
\description{
Match a selector node to a container
}
\keyword{internal}
