% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{parse_selector}
\alias{parse_selector}
\title{Parse a CSS3-like selector}
\usage{
parse_selector(x)
}
\arguments{
\item{x}{CSS3-like selector string.}
}
\description{
This is the rather small subset of CSS3 that is supported:
}
\details{
Selectors:
\itemize{
\item Type selectors, e.g. \code{input} selects all \code{<input>} elements.
\item Class selectors, e.g. \code{.index} selects any element that has a class
of "index".
\item ID selector. \code{#toc} will match the element that has the ID "toc".
}

Combinators:
\itemize{
\item Descendant combinator, i.e. the space, that combinator selects nodes
that are descendants of the first element. E.g. \code{div span} will match
all \code{<span>} elements that are inside a \code{<div>} element.
}
}
\keyword{internal}
