% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluralize.R
\name{no}
\alias{no}
\alias{qty}
\title{Pluralization helper functions}
\usage{
no(expr)

qty(expr)
}
\arguments{
\item{expr}{For \code{no()} it is an expression that is printed as "no" in
cli expressions, it is interpreted as a zero quantity. For \code{qty()}
an expression that sets the pluralization quantity without printing
anything. See examples below.}
}
\description{
Pluralization helper functions
}
\seealso{
Other pluralization: 
\code{\link{pluralization}}
}
\concept{pluralization}
