% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettycode.R
\name{pretty_print_code}
\alias{pretty_print_code}
\title{Turn on pretty-printing functions at the R console}
\usage{
pretty_print_code()
}
\description{
Defines a print method for functions, in the current session, that supports
syntax highlighting.
}
\details{
The new print method takes priority over the built-in one. Use
\code{\link[base:message]{base::suppressMessages()}} to suppress the alert message.
}
