% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-utils.R
\name{cli_progress_num}
\alias{cli_progress_num}
\alias{cli_progress_cleanup}
\title{Progress bar utility functions.}
\usage{
cli_progress_num()

cli_progress_cleanup()
}
\value{
\code{cli_progress_num()} returns an integer scalar.

`cli_progress_cleanup() does not return anything.
}
\description{
Progress bar utility functions.
}
\details{
\code{cli_progress_num()} returns the number of currently
active progress bars. (These do not currently include the progress
bars created in C/C++ code.)

\code{cli_progress_cleanup()} terminates all active progress bars.
(It currently ignores progress bars created in the C/C++ code.)
}
\seealso{
Other progress bar functions: 
\code{\link{cli_progress_along}()},
\code{\link{cli_progress_bar}()},
\code{\link{cli_progress_builtin_handlers}()},
\code{\link{cli_progress_message}()},
\code{\link{cli_progress_output}()},
\code{\link{cli_progress_step}()},
\code{\link{cli_progress_styles}()},
\code{\link{progress-variables}}
}
\concept{progress bar functions}
