% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\docType{data}
\name{should_tick}
\alias{should_tick}
\alias{cli_tick_reset}
\alias{ccli_tick_reset}
\title{Whether it is time to update the cli status bar}
\format{
An object of class \code{logical} of length 1.

An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
should_tick

cli_tick_reset()

ccli_tick_reset
}
\description{
TODO
}
\keyword{datasets}
