% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{spark_line}
\alias{spark_line}
\title{Draw a sparkline line graph with Braille characters.}
\usage{
spark_line(x)
}
\arguments{
\item{x}{A numeric vector between 0 and 1}
}
\value{
A character string of class \code{cli_spark_line} and \code{cli_spark}.
}
\description{
You might want to avoid sparklines on non-UTF-8 systems, because they
do not look good. You can use \code{\link[=is_utf8_output]{is_utf8_output()}} to test for support
for them.
}
\details{
\if{html}{\out{<div class="sourceCode r">}}\preformatted{x <- seq(0, 1, length = 10)
spark_line(x)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> ⣀⡠⠔⠊⠉                                                                           
</pre></div>
}}
}
\seealso{
\code{\link[=spark_bar]{spark_bar()}}
}
