% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-component-tree.R
\name{style_component_tree}
\alias{style_component_tree}
\title{Style a component tree: inherit styles to child nodes}
\usage{
style_component_tree(node)
}
\arguments{
\item{node}{A themed component tree, the outout of
\code{\link[=theme_component_tree]{theme_component_tree()}}.}
}
\value{
A styled component tree.
}
\description{
Styling a component tree is part of the
\link[=component-trees]{rendering process}.
}
\details{
We walk the component tree and inheerit styles from parent nodes to
their child nodes, according to the style inhertance rules.
This is performed by the \code{\link[=inherit_styles]{inherit_styles()}} helper function, and each
style may have its own inheritance rules.

Text pieces are special, because they cannot be themed, so they don't
have any style \emph{before} this function.
But they can have styles \emph{after} this function, if styles are inherited
from parant elements.
}
\seealso{
Other component trees: 
\code{\link{component-trees}},
\code{\link{map_component_tree}()},
\code{\link{theme_component_tree}()}
}
\concept{component trees}
\keyword{internal}
