% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-component-tree.R
\name{theme_component_tree}
\alias{theme_component_tree}
\title{Theme a component tree}
\usage{
theme_component_tree(node, theme = list())
}
\arguments{
\item{node}{A (mapped) component tree, the output of
\code{\link[=map_component_tree]{map_component_tree()}}.}

\item{theme}{Theme to apply to the tree. Note that this internal
function does \strong{not} use the \code{cli.theme} or \code{cli.user_theme}
options of the built-in theme. It only uses the \code{theme} argument.}
}
\value{
A themed component tree.
}
\description{
Theming a component tree is part of the
\link[=component-trees]{rendering process}.
}
\details{
For each node in the tree, we find the matching selectors in the theme,
and add the styles of the matching selectors to the node.
The style of the node (that was specified when the corrsponding
component was created) takes precedence over the theme, see
\code{\link[=apply_theme_to_style]{apply_theme_to_style()}}.
}
\seealso{
Other component trees: 
\code{\link{component-trees}},
\code{\link{map_component_tree}()},
\code{\link{style_component_tree}()}
}
\concept{component trees}
\keyword{internal}
