% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-tick-along.R
\name{tick_along}
\alias{tick_along}
\title{Add a progress bar to a mapping function or for loop}
\usage{
tick_along(x, name = NULL, total = length(x), ..., .envir = parent.frame())
}
\arguments{
\item{x}{Sequence to add the progress bar to.}

\item{name}{Name of the progtess bar, a label, passed to
\code{\link[=cli_progress_bar]{cli_progress_bar()}}.}

\item{total}{Passed to \code{\link[=cli_progress_bar]{cli_progress_bar()}}.}

\item{...}{Passed to \code{\link[=cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[=cli_progress_bar]{cli_progress_bar()}}.}
}
\value{
An index vector from 1 to \code{length(x)} that triggers progress
updates as you iterate over it.
}
\description{
Use \code{tick_along()} in a mapping function or in a for loop, to add a
progress bar. It uses \code{\link[=cli_progress_bar]{cli_progress_bar()}} internally.
}
\details{
Usage:\if{html}{\out{<div class="sourceCode r">}}\preformatted{lapply(tick_along(X), function(i) ...)
}\if{html}{\out{</div>}}\if{html}{\out{<div class="sourceCode r">}}\preformatted{for (i in tick_along(seq)) \{
  ...
\}
}\if{html}{\out{</div>}}

Note that if you use \code{break} in the \code{for} loop, you probably want to
terminate the progress bar explicitly when breaking out of the loop,
or right after the loop:\if{html}{\out{<div class="sourceCode r">}}\preformatted{for (i in tick_along(seq)) \{
  ...
  if (cond) cli_progress_done() && break
  ...
\}
}\if{html}{\out{</div>}}
}
\seealso{
\code{\link[=cli_progress_bar]{cli_progress_bar()}} and the traditional progress bar API.
\code{\link[=ticking]{ticking()}} to add a progress bar to a \code{while} loop.
}
