% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-ticking.R
\name{ticking}
\alias{ticking}
\title{Add a progress bar to a \code{while} loop}
\usage{
ticking(cond, name = NULL, ..., .envir = parent.frame())
}
\arguments{
\item{cond}{Loop condition. Once this evaluates to \code{FALSE} the
progress bar will be terminated.}

\item{name}{Name of the progress bar, a label, passed to
\code{\link[=cli_progress_bar]{cli_progress_bar()}}.}

\item{...}{Passed to \code{\link[=cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[=cli_progress_bar]{cli_progress_bar()}}.}
}
\value{
The value of the \code{cond} condition.
}
\description{
Note that this function is currently experimental!

Use \code{ticking()} in the condition of a while loop, to add a progress bar.
It uses \code{\link[=cli_progress_bar]{cli_progress_bar()}} internally.
}
\details{
Usage:\if{html}{\out{<div class="sourceCode r">}}\preformatted{while (ticking(cond)) \{
  ...
\}
}\if{html}{\out{</div>}}

Note that if you use \code{break} in the loop, you probably want to terminate
the progress bar explicitly when breaking out or right after the loop:\if{html}{\out{<div class="sourceCode r">}}\preformatted{while (ticking(cond)) \{
  ...
  if (cond) cli_progress_done() && break
  ...
\}
}\if{html}{\out{</div>}}
}
