% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tty.R
\name{ansi_hide_cursor}
\alias{ansi_hide_cursor}
\alias{ansi_show_cursor}
\alias{ansi_with_hidden_cursor}
\title{Hide/show cursor in a terminal}
\usage{
ansi_hide_cursor(stream = stderr())

ansi_show_cursor(stream = stderr())

ansi_with_hidden_cursor(expr, stream = stderr())
}
\arguments{
\item{stream}{The stream of the terminal to output the ANSI sequence to.}

\item{expr}{R expression to evaluate.}
}
\description{
This only works in terminal emulators. In other environments, it
does nothing.
}
\details{
\code{ansi_hide_cursor()} hides the cursor.

\code{ansi_show_cursor()} shows the cursor.

\code{ansi_with_hidden_cursor()} temporarily hides the cursor for
evaluating an expression.
}
\concept{terminal capabiltiies}
