% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansiex.R
\name{ansi_nzchar}
\alias{ansi_nzchar}
\title{Like \code{\link[base:nchar]{base::nzchar()}}, but for ANSI strings}
\usage{
ansi_nzchar(x, ...)
}
\arguments{
\item{x}{Charcater vector. Other objects are coarced using
\code{\link[base:character]{base::as.character()}}.}

\item{...}{Passed to \code{\link[base:nchar]{base::nzchar()}}.}
}
\description{
Like \code{\link[base:nchar]{base::nzchar()}}, but for ANSI strings
}
\examples{
ansi_nzchar("")
ansi_nzchar(col_red(""))
}
