% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crayonex.R
\name{ansi_strip}
\alias{ansi_strip}
\title{Remove ANSI escape sequences from a string}
\usage{
ansi_strip(string)
}
\arguments{
\item{string}{The input string.}
}
\value{
The cleaned up string.
}
\description{
The input may be of class \code{ansi_string} class, this is also dropped
from the result.
}
\examples{
ansi_strip(col_red("foobar")) == "foobar"
}
\seealso{
Other low level ANSI functions: 
\code{\link{ansi_has_any}()},
\code{\link{ansi_hide_cursor}()},
\code{\link{ansi_regex}()}
}
\concept{low level ANSI functions}
