% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\docType{data}
\name{cli}
\alias{cli}
\title{The default command line interface instant}
\format{An object of class \code{cli_class} (inherits from \code{R6}) of length 25.}
\usage{
cli
}
\description{
TODO
}
\keyword{datasets}
