% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_h1}
\alias{cli_h1}
\alias{cli_h2}
\alias{cli_h3}
\title{CLI headers}
\usage{
cli_h1(text, id = NULL, class = NULL, .envir = parent.frame())

cli_h2(text, id = NULL, class = NULL, .envir = parent.frame())

cli_h3(text, id = NULL, class = NULL, .envir = parent.frame())
}
\arguments{
\item{text}{Text of the header. It can contain inline markup.}

\item{id}{Id of the header element, string. It can be used in themes.}

\item{class}{Class of the header element,  string. It can be used in
themes.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
CLI headers
}
\examples{
cli_h1("Main title")
cli_h2("Subtitle")
cli_text("And some regular text....")
}
