% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progress-bar.R
\name{cli_progress_styles}
\alias{cli_progress_styles}
\title{List of built-in cli progress styles}
\usage{
cli_progress_styles()
}
\value{
A named list with sublists containing elements
\code{complete}, \code{incomplete} and potentially \code{current}.
}
\description{
List of built-in cli progress styles
}
