% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-bar.R
\name{cli_status}
\alias{cli_status}
\title{Update the status bar}
\usage{
cli_status(..., .keep = FALSE, .auto_close = TRUE, .envir = parent.frame())
}
\arguments{
\item{...}{The text to show, in character vectors. They will be
concatenated into a single string. Long lines are cut to
\code{\link[=console_width]{console_width()}}.}

\item{.keep}{What to do when this status bar is cleared. If \code{TRUE} then
the content of this status bar is kept, as regular cli output (the
screen is scrolled up if needed). If \code{FALSE}, then this status bar
is deleted.}

\item{.auto_close}{Whether to clear the status bar when the calling
function finishes (or ‘.envir’ is removed from the stack, if
specified).}

\item{.envir}{Environment to evaluate the glue expressions in. It is
also used to auto-clear the status bar if ‘.auto_close’ is ‘TRUE’.}
}
\value{
The id of the new status bar container element, invisibly.
}
\description{
The status bar is the last line of the terminal. cli apps can use this
to show status information, progress bars, etc. The status bar is kept
intact by all semantic cli output.
}
\details{
Use \code{\link[=cli_status_clear]{cli_status_clear()}} to clear the status bar.
}
\seealso{
Other status bar: 
\code{\link{cli_status_clear}()},
\code{\link{cli_status_update}()}
}
\concept{status bar}
