% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status-bar.R
\name{cli_status_clear}
\alias{cli_status_clear}
\title{Clear the status bar}
\usage{
cli_status_clear(id = NULL)
}
\arguments{
\item{id}{Id of the status bar container to clear. If \code{id} is not the id
of the current status bar (because it was overwritten by another
status bar container), then the status bar is not cleared. If \code{NULL}
(the default) then the status bar is always cleared.}
}
\description{
Clear the status bar
}
\seealso{
Other status bar: \code{\link{cli_status_update}},
  \code{\link{cli_status}}
}
\concept{status bar}
