% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{cli_verbatim}
\alias{cli_verbatim}
\title{CLI verbatim text}
\usage{
cli_verbatim(..., .envir = parent.frame())
}
\arguments{
\item{...}{The text to show, in character vectors. Each element is
printed on a new line.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
It is not wrapped, but printed as is.
}
\examples{
cli_verbatim("This has\nthree", "lines")
}
