% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term-width.R, R/width.R
\name{console_width}
\alias{console_width}
\alias{console_width}
\title{Determine the width of the console}
\usage{
console_width()

console_width()
}
\value{
Integer scalar, the console with, in number of characters.

Integer scalar, the console with, in number of characters.
}
\description{
It uses the \code{RSTUDIO_CONSOLE_WIDTH} environment variable, if set.
Otherwise it uses the \code{width} option. If this is not set either,
then 80 is used.

It uses the \code{RSTUDIO_CONSOLE_WIDTH} environment variable, if set.
Otherwise it uses the \code{width} option. If this is not set either,
then 80 is used.
}
