% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{ediff_str}
\alias{ediff_str}
\title{Compare two character strings, character by character}
\usage{
ediff_str(old, new)
}
\arguments{
\item{old}{First string, must not be \code{NA}.}

\item{new}{Second string, must not be \code{NA}.}
}
\description{
Characters are defined by UTF-8 graphemes.
}
\examples{
str1 <- "abcdefghijklmnopqrstuvwxyz"
str2 <- "PREabcdefgMIDDLEnopqrstuvwxyzPOST"
ediff_str(str1, str2)
}
\seealso{
The diffobj package for a much more comprehensive set of
\code{diff}-like tools.

Other diff functions in cli: 
\code{\link{ediff_chr}()}
}
\concept{diff functions in cli}
