% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_inline}
\alias{format_inline}
\title{Format and returns a line of text}
\usage{
format_inline(...)
}
\arguments{
\item{...}{Passed to \code{\link[=cli_text]{cli_text()}}.}
}
\value{
Character scalar, the formatted string.
}
\description{
You can use this function to format a line of cli text, without emitting
it to the screen. It uses \code{\link[=cli_text]{cli_text()}} internally.
}
\examples{
format_inline("This is a message for {.emph later}.")
}
