% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{format_inline}
\alias{format_inline}
\title{Format and returns a line of text}
\usage{
format_inline(..., .envir = parent.frame(), collapse = TRUE)
}
\arguments{
\item{...}{Passed to \code{\link[=cli_text]{cli_text()}}.}

\item{.envir}{Environment to evaluate the expressions in.}

\item{collapse}{Whether to collapse the result if it has multiple
lines, e.g. because of \verb{\\f} characters.}
}
\value{
Character scalar, the formatted string.
}
\description{
You can use this function to format a line of cli text, without emitting
it to the screen. It uses \code{\link[=cli_text]{cli_text()}} internally.
}
\details{
\code{format_inline()} performs no width-wrapping.
}
\examples{
format_inline("A message for {.emph later}, thanks {.fn format_inline}.")
}
