% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cliapp-docs.R
\name{links}
\alias{links}
\title{cli hyperlinks}
\description{
cli hyperlinks
}
\section{About ANSI hyperlinks}{


Certain cli styles create clickable links, if your IDE or terminal
supports them.
}

\section{Hyperlink Support}{


As of today, RStudio (from version v2022.07.0+548)
supports all hyperlink types discussed here. Certain terminals, e.g.
iTerm on macOS, Linux terminals based on VTE (GNOME terminal) support
\code{.href}, \code{.email} and \code{.file} links.

You can use \code{\link[=ansi_has_hyperlink_support]{ansi_has_hyperlink_support()}} to check if your terminal or
IDE has hyperlink support in general, and \code{\link[=ansi_hyperlink_types]{ansi_hyperlink_types()}} to
check if various types of hyperlinks are supported.

If your hyperlink support is not detected properly in your IDE or
terminal, please open a cli issue at
\url{https://github.com/r-lib/cli/issues}.
}

\section{URLs}{


There are two cli styles to link to generic URLs. \code{.url} does not
allow custom link text, but \verb{\\href} does.

Examples:

\if{html}{\out{<div class="sourceCode r">}}\preformatted{cli_text(
  "See the cli homepage at \{.url https://cli.r-lib.org\} for details."
)
}\if{html}{\out{</div>}}\if{html}{\out{
<div class="asciicast" style="color: #172431;font-family: 'Fira Code',Monaco,Consolas,Menlo,'Bitstream Vera Sans Mono','Powerline Symbols',monospace;line-height: 1.300000"><pre>
#> See the cli homepage at <span style="font-style: italic;;color: #268BD2">&lt;https://cli.r-lib.org&gt;</span> for details.                    
</pre></div>
}}
}

\section{Files}{

}

\section{Functions, help topics and vignettes}{

}

\section{Click to run code}{

}

\section{Link text}{

}

\section{Vectorization}{

}

\section{Custom links}{

}

