% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rules.R
\name{rule}
\alias{rule}
\title{Make a rule with one or two text labels}
\usage{
rule(left = "", center = "", right = "", line = 1, line_color = NULL,
  width = getOption("width"))
}
\arguments{
\item{left}{Label to show on the left. It interferes with the \code{center}
label, only at most one of them can be present.}

\item{center}{Label to show at the center. It interferes  with the
\code{left} and \code{right} labels.}

\item{right}{Label to show on the right. It interferes with the \code{center}
label, only at most one of them can be present.}

\item{line}{The character or string that is used to draw the line.
It can also \code{1} or \code{2}, to request a single line (Unicode, if
available), or a double line.}

\item{line_color}{Either a color name (used in \code{\link[crayon:make_style]{crayon::make_style()}}),
or a style function from \code{crayon}, to color the line.}

\item{width}{Width of the rule. Defaults to the \code{width} option, see
\code{\link[base:options]{base::options()}}.}
}
\value{
Character scalar, the rule.
}
\description{
The rule can include either a centered text label, or labels on the
left and right side.
}
\details{
To color the labels, use the functions from the \code{crayon} package
directly, see examples below. To color the line, either use the
\code{crayon} colors directly, or the \code{line_color} option.
}
\examples{

## Simple rule
cat(rule(), "\\n")

## Double rule
cat(rule(line = 2), "\\n")

## Left label
cat(rule(left = "Results"), "\\n")

## Centered label
cat(rule(center = " * RESULTS * "), "\\n")

## Colored labels
cat(rule(center = crayon::red(" * RESULTS * ")), "\\n")

## Colored line
cat(rule(center = crayon::red(" * RESULTS * "), line_color = "red"), "\\n")

## Custom line
cat(rule(center = "TITLE", line = "~"))

## More custom line
cat(rule(center = "TITLE", line = crayon::blue("~-")), "\\n")

## Even more custom line
cat(rule(center = crayon::bgRed(" ", clisymbols::symbol$star, "TITLE",
  clisymbols::symbol$star, " "),
  line = "\\u2582",
  line_color = "orange"))
}
