% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ruler.R
\name{ruler}
\alias{ruler}
\title{Print the helpful ruler to the screen}
\usage{
ruler(width = console_width())
}
\arguments{
\item{width}{Ruler width.}
}
\description{
Print the helpful ruler to the screen
}
\examples{
ruler()
}
