% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\docType{data}
\name{should_tick}
\alias{should_tick}
\alias{cli_tick_reset}
\alias{ccli_tick_reset}
\title{The cli timer}
\usage{
should_tick

cli_tick_reset()

ccli_tick_reset
}
\value{
\code{should_tick} is a logical scalar on R vesions 3.6.0 and later,
it is an integer scalar with values 0L or 1L on R 3.5.x. It is a logical
scalar on older r vresions.

\code{cli_tick_reset()} returns \code{NULL}.
}
\description{
cli includes a timer that is alerted periodically. When the the
timer is alerted, the \code{should_tick} variable is set to \code{TRUE}.

\code{cli_tick_reset()} resets it to \code{FALSE}.

\code{ccli_tick_reset} can be used in \code{.Call()} and it is marginally faster
than calling `cli_tick_reset().

By default the timer is alerted every 200
milliseconds in interactive sessions and every
3000 milliseconds in non-interactive
sessions.
}
\keyword{datasets}
