% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\docType{data}
\name{cli_tick_reset}
\alias{cli_tick_reset}
\alias{ccli_tick_reset}
\title{\code{cli_tick_reset()} resets the cli status bar timer after an update}
\format{
An object of class \code{CallRoutine} (inherits from \code{NativeSymbolInfo}) of length 4.
}
\usage{
cli_tick_reset()

ccli_tick_reset
}
\description{
\code{cli_tick_reset()} resets the cli status bar timer after an update
}
\keyword{datasets}
