% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{start_app}
\alias{start_app}
\alias{stop_app}
\alias{default_app}
\title{Start, stop, query the default cli application}
\usage{
start_app(theme = getOption("cli.theme"), output = c("message",
  "stdout"), .auto_close = TRUE, .envir = parent.frame())

stop_app(app = NULL)

default_app()
}
\arguments{
\item{theme}{Theme to use.}

\item{output}{How to print the output.}

\item{.auto_close}{Whether to stop the app, when the calling frame
is destroyed.}

\item{.envir}{The environment to use, instead of the calling frame,
to trigger the stop of the app.}

\item{app}{App to stop. If \code{NULL}, the current default app is stopped.
Otherwise we find the supplied app in the app stack, and remote it,
together with all the apps above it.}
}
\value{
\code{start_app} returns the new app, \code{default_app} returns the default app.
\code{stop_app} does not return anything.
}
\description{
\code{start_app} creates an app, and places it on the top of the app stack.
}
\details{
\code{stop_app} removes the top app, or multiple apps from the app stack.

\code{default_app} returns the default app, the one on the top of the stack.
}
