% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansi-hyperlink.R
\name{style_hyperlink}
\alias{style_hyperlink}
\alias{ansi_has_hyperlink_support}
\title{Terminal Hyperlinks}
\usage{
style_hyperlink(text, url)

ansi_has_hyperlink_support()
}
\arguments{
\item{text}{Text to show. \code{text} and \code{url} are recycled to match their
length, via a \code{paste0()} call.}

\item{url}{URL to link to.}
}
\value{
Styled \code{ansi_string} for \code{style_hyperlink()}.
Logical scalar for \code{ansi_has_hyperlink_support()}.
}
\description{
Terminal Hyperlinks
}
\details{
\code{ansi_hyperlink()} creates an ANSI hyperlink.

\code{ansi_has_hyperlink_support()} checks if the current \code{stdout()}
supports hyperlinks.

See also
\url{https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda}.
}
\examples{
cat("This is an", style_hyperlink("R", "https://r-project.org"), "link.\n")
ansi_has_hyperlink_support()
}
