% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol.R
\name{symbol}
\alias{symbol}
\title{Various handy symbols to use in a command line UI}
\format{A named list, see \code{names(symbol)} for all sign names.}
\usage{
symbol
}
\description{
Various handy symbols to use in a command line UI
}
\details{
On Windows they have a fallback to less fancy symbols.
}
\examples{
cat(symbol$check, " SUCCESS\\n", symbol$cross, " FAILURE\\n", sep="")

## All symbols
cat(paste(format(names(symbol), width=20),
  unlist(symbol)), sep = "\\n")
}
